requirements.txt
djangorestframework-simplejwt
django-elasticsearch-dsl
simple-file-checksum
django-cors-headers
djangorestframework
django-prometheus
django-ratelimit
confluent-kafka
psycopg2-binary
django-environ
python-dotenv
django-filter
django_redis
sentry_sdk
jdatetime
injector
pydantic
watchdog
coverage
requests
gunicorn
celery
Pillow
django
redis
faker
minio
pike
pytz

